improve the pattern
